test: split contest tier coverage
